raomve rates request so App will not reload
