+++ src/containers/Filters/Filters.tsx
@@ -3,11 +3,10 @@
 /* eslint-disable react/jsx-wrap-multilines */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable object-curly-newline */
-import React, { RefObject, useCallback, useMemo, useState, VFC } from 'react';
+import React, { RefObject, useCallback, useEffect, useMemo, useState, VFC } from 'react';
 import cn from 'classnames';
 import arrowDown from 'assets/img/icons/arrowDown.svg';
 import arrowUp from 'assets/img/icons/arrowUp.svg';
-import { matic, usdt } from 'assets/img';
 import { validateOnlyNumbers } from 'utils';
 import {
   Button,
@@ -21,6 +20,12 @@ import {
 import { CloseIcon } from 'assets/icons';
 import BigNumber from 'bignumber.js';
 import { Collection } from 'types/api';
+import { useShallowSelector } from 'hooks';
+import ratesSelector from 'store/rates/selectors';
+import userSelector from 'store/user/selectors';
+import { useDispatch } from 'react-redux';
+import { getRates } from 'store/rates/actions';
+import { Chains } from 'types';
 import styles from './styles.module.scss';
 
 export interface FiltersProps {
@@ -42,21 +47,6 @@ export interface FiltersProps {
   onLoadMore?: (page: number) => void
 }
 
-export const rates = [
-  {
-    rate: '425.000000000',
-    symbol: 'USDT',
-    name: 'Usdt',
-    image: usdt,
-  },
-  {
-    rate: '123.000000000',
-    symbol: 'MATIC',
-    name: 'Matic',
-    image: matic,
-  },
-];
-
 export const Filters: VFC<FiltersProps> = ({
   filters,
   onClose,
@@ -75,6 +65,9 @@ export const Filters: VFC<FiltersProps> = ({
   totalCollectionsPages,
   onLoadMore = () => {},
 }) => {
+  const rates = useShallowSelector(ratesSelector.getProp('rates'));
+  const dispatch = useDispatch();
+  const userNetwork = useShallowSelector(userSelector.getProp('chain'));
   const [minValue, setMinValue] = useState('');
   const [maxValue, setMaxValue] = useState('');
   const { standart, isAuction, currency, collections, orderBy } = filters;
@@ -83,6 +76,13 @@ export const Filters: VFC<FiltersProps> = ({
       minValue && maxValue && !new BigNumber(minValue).isLessThanOrEqualTo(new BigNumber(maxValue)),
     [maxValue, minValue],
   );
+  const handleGetRates = useCallback(() => {
+    dispatch(getRates({ network: userNetwork || Chains.polygon }));
+  }, [dispatch, userNetwork]);
+
+  useEffect(() => {
+    handleGetRates();
+  }, [handleGetRates]);
 
   const handleChangeCurrencyValue = useCallback(
     (newCurrency) => {
@@ -164,7 +164,7 @@ export const Filters: VFC<FiltersProps> = ({
     setMinValue('');
   }, [handleClearFilters]);
 
-  const currencyOptions = [
+  const currencyOptions = useMemo(() => (rates.length ? [
     ...rates.map((rate, index) => ({
       id: String(index),
       content: (
@@ -173,7 +173,7 @@ export const Filters: VFC<FiltersProps> = ({
           onChange={() => handleChangeCurrencyValue(rate.symbol)}
           content={
             <div className={styles.currency}>
-              <img src={rate.image} alt="currency" />
+              <img src={rate.image} alt="currency" className={styles.rate} />
               <Text variant="body-2" color="light1">
                 {rate.symbol}
               </Text>
@@ -182,7 +182,7 @@ export const Filters: VFC<FiltersProps> = ({
         />
       ),
     })),
-  ];
+  ] : []), [currency, handleChangeCurrencyValue, rates]);
   const dateOptions = [
     {
       id: '0',
